+++ aspnet-core/src/BookingWeb.Core/DbEntities/Phong.cs
@@ -12,10 +12,9 @@ namespace BookingWeb.DbEntities
 
         public string TenFileAnhDaiDien { get; set; }
 
-        public float DiemDanhGiaTB { get; set; }
+        public double DiemDanhGiaTB { get; set; }
 
-        public float DanhGiaSaoTb { get; set; }
-
+        public double DanhGiaSaoTb { get; set; }
 
         public int? DonViKinhDoanhId { get; set; }
